ЗАменены анимации лотти. Навел порядок в папках. Удалил ненужные файлы.
